Refactor: Pass caching strategy as parameter to IonClient instance
